Make CachingHostAllocator work with memory pools.
Only allocation to a cuda graph's private pool via stream capture is
supported.

Allocation to a memory pool in non-stream-captured code is not
supported. There is no obvious usecase at this time for that.

In stream capture, we refuse to reuse a host memory block as soon as
record_event() is called on that block. This is to prevent a
stream-captured CUDA kernel from reading different contents from a
memory block than would be read if, counterfactually, this CUDA
kernels were running eagerly on a cuda stream.

See
https://github.com/pytorch/pytorch/pull/161583#issuecomment-3229885771
for elaboration.

ghstack-source-id: c7c2f4682708e38ed3cdca5c7e6830b74e612afa
Pull-Request: https://github.com/pytorch/pytorch/pull/167507